move InstructionInserter to instruction.rs
