+++ sway-ir/src/block.rs
@@ -11,17 +11,15 @@
 //! Every [`Function`] has at least one block, the first of which is usually labeled `entry`.
 
 use rustc_hash::{FxHashMap, FxHashSet};
-use sway_types::Ident;
 
 use crate::{
     context::Context,
     error::IrError,
     function::Function,
     instruction::{FuelVmInstruction, InstOp},
-    pretty::DebugWithContext,
     value::{Value, ValueDatum},
-    AsmArg, AsmBlock, AsmInstruction, BinaryOpKind, BranchToWithArgs, Constant, Instruction,
-    LocalVar, Predicate, Register, Type, UnaryOpKind,
+    BranchToWithArgs, DebugWithContext, Instruction, InstructionInserter, InstructionIterator,
+    Type,
 };
 
 /// A wrapper around an [ECS](https://github.com/fitzgen/generational-arena) handle into the
@@ -36,7 +34,7 @@ pub struct BlockContent {
     /// The function containing this block.
     pub function: Function,
     /// List of instructions in the block.
-    instructions: Vec<Value>,
+    pub(crate) instructions: Vec<Value>,
     /// Block arguments: Another form of SSA PHIs.
     pub args: Vec<Value>,
     /// CFG predecessors
@@ -595,484 +593,5 @@ impl Iterator for BlockIterator {
         } else {
             None
         }
-    }
-}
-
-/// Iterate over all [`Instruction`]s in a specific [`Block`].
-pub struct InstructionIterator {
-    instructions: Vec<generational_arena::Index>,
-    next: usize,
-    next_back: isize,
-}
-
-impl InstructionIterator {
-    pub fn new(context: &Context, block: &Block) -> Self {
-        // Copy all the current instruction indices, so they may be modified in the context during
-        // iteration.
-        InstructionIterator {
-            instructions: context.blocks[block.0]
-                .instructions
-                .iter()
-                .map(|val| val.0)
-                .collect(),
-            next: 0,
-            next_back: context.blocks[block.0].instructions.len() as isize - 1,
-        }
-    }
-}
-
-impl Iterator for InstructionIterator {
-    type Item = Value;
-
-    fn next(&mut self) -> Option<Value> {
-        if self.next < self.instructions.len() {
-            let idx = self.next;
-            self.next += 1;
-            Some(Value(self.instructions[idx]))
-        } else {
-            None
-        }
-    }
-}
-
-impl DoubleEndedIterator for InstructionIterator {
-    fn next_back(&mut self) -> Option<Value> {
-        if self.next_back >= 0 {
-            let idx = self.next_back;
-            self.next_back -= 1;
-            Some(Value(self.instructions[idx as usize]))
-        } else {
-            None
-        }
-    }
-}
-
-/// Where to insert new instructions in the block.
-pub enum InsertionPosition {
-    // Insert at the start of the basic block.
-    Start,
-    // Insert at the end of the basic block (append).
-    End,
-    // Insert after instruction.
-    After(Value),
-    // Insert before instruction.
-    Before(Value),
-    // Insert at position / index.
-    At(usize),
-}
-
-/// Provide a context for inserting new [`Instruction`]s to a [`Block`].
-pub struct InstructionInserter<'a, 'eng> {
-    context: &'a mut Context<'eng>,
-    block: Block,
-    position: InsertionPosition,
-}
-
-macro_rules! insert_instruction {
-    ($self: ident, $ctor: expr) => {{
-        let instruction_val = Value::new_instruction($self.context, $self.block, $ctor);
-        let pos = $self.get_position_index();
-        let instructions = &mut $self.context.blocks[$self.block.0].instructions;
-        instructions.insert(pos, instruction_val);
-        instruction_val
-    }};
-}
-
-impl<'a, 'eng> InstructionInserter<'a, 'eng> {
-    /// Return a new [`InstructionInserter`] context for `block`.
-    pub fn new(
-        context: &'a mut Context<'eng>,
-        block: Block,
-        position: InsertionPosition,
-    ) -> InstructionInserter<'a, 'eng> {
-        InstructionInserter {
-            context,
-            block,
-            position,
-        }
-    }
-
-    // Recomputes the index in the instruction vec. O(n) in the worst case.
-    fn get_position_index(&self) -> usize {
-        let instructions = &self.context.blocks[self.block.0].instructions;
-        match self.position {
-            InsertionPosition::Start => 0,
-            InsertionPosition::End => instructions.len(),
-            InsertionPosition::After(inst) => {
-                instructions
-                    .iter()
-                    .position(|val| *val == inst)
-                    .expect("Provided position for insertion does not exist")
-                    + 1
-            }
-            InsertionPosition::Before(inst) => instructions
-                .iter()
-                .position(|val| *val == inst)
-                .expect("Provided position for insertion does not exist"),
-            InsertionPosition::At(pos) => pos,
-        }
-    }
-
-    // Insert a slice of instructions.
-    pub fn insert_slice(&mut self, slice: &[Value]) {
-        let pos = self.get_position_index();
-        self.context.blocks[self.block.0]
-            .instructions
-            .splice(pos..pos, slice.iter().cloned());
-    }
-
-    // Insert a single instruction.
-    pub fn insert(&mut self, inst: Value) {
-        let pos = self.get_position_index();
-        self.context.blocks[self.block.0]
-            .instructions
-            .insert(pos, inst);
-    }
-
-    //
-    // XXX Maybe these should return result, in case they get bad args?
-    //
-
-    /// Append a new [`Instruction::AsmBlock`] from `args` and a `body`.
-    pub fn asm_block(
-        self,
-        args: Vec<AsmArg>,
-        body: Vec<AsmInstruction>,
-        return_type: Type,
-        return_name: Option<Ident>,
-    ) -> Value {
-        let asm = AsmBlock::new(
-            args.iter().map(|arg| arg.name.clone()).collect(),
-            body,
-            return_type,
-            return_name,
-        );
-        self.asm_block_from_asm(asm, args)
-    }
-
-    pub fn asm_block_from_asm(self, asm: AsmBlock, args: Vec<AsmArg>) -> Value {
-        insert_instruction!(self, InstOp::AsmBlock(asm, args))
-    }
-
-    pub fn bitcast(self, value: Value, ty: Type) -> Value {
-        insert_instruction!(self, InstOp::BitCast(value, ty))
-    }
-
-    pub fn unary_op(self, op: UnaryOpKind, arg: Value) -> Value {
-        insert_instruction!(self, InstOp::UnaryOp { op, arg })
-    }
-
-    pub fn wide_unary_op(self, op: UnaryOpKind, arg: Value, result: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::WideUnaryOp { op, arg, result })
-        )
-    }
-
-    pub fn wide_binary_op(
-        self,
-        op: BinaryOpKind,
-        arg1: Value,
-        arg2: Value,
-        result: Value,
-    ) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::WideBinaryOp {
-                op,
-                arg1,
-                arg2,
-                result
-            })
-        )
-    }
-
-    pub fn wide_modular_op(
-        self,
-        op: BinaryOpKind,
-        result: Value,
-        arg1: Value,
-        arg2: Value,
-        arg3: Value,
-    ) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::WideModularOp {
-                op,
-                result,
-                arg1,
-                arg2,
-                arg3,
-            })
-        )
-    }
-
-    pub fn wide_cmp_op(self, op: Predicate, arg1: Value, arg2: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::WideCmpOp { op, arg1, arg2 })
-        )
-    }
-
-    pub fn binary_op(self, op: BinaryOpKind, arg1: Value, arg2: Value) -> Value {
-        insert_instruction!(self, InstOp::BinaryOp { op, arg1, arg2 })
-    }
-
-    pub fn branch(self, to_block: Block, dest_params: Vec<Value>) -> Value {
-        let br_val = Value::new_instruction(
-            self.context,
-            self.block,
-            InstOp::Branch(BranchToWithArgs {
-                block: to_block,
-                args: dest_params,
-            }),
-        );
-        to_block.add_pred(self.context, &self.block);
-        self.context.blocks[self.block.0].instructions.push(br_val);
-        br_val
-    }
-
-    pub fn call(self, function: Function, args: &[Value]) -> Value {
-        insert_instruction!(self, InstOp::Call(function, args.to_vec()))
-    }
-
-    pub fn cast_ptr(self, val: Value, ty: Type) -> Value {
-        insert_instruction!(self, InstOp::CastPtr(val, ty))
-    }
-
-    pub fn cmp(self, pred: Predicate, lhs_value: Value, rhs_value: Value) -> Value {
-        insert_instruction!(self, InstOp::Cmp(pred, lhs_value, rhs_value))
-    }
-
-    pub fn conditional_branch(
-        self,
-        cond_value: Value,
-        true_block: Block,
-        false_block: Block,
-        true_dest_params: Vec<Value>,
-        false_dest_params: Vec<Value>,
-    ) -> Value {
-        let cbr_val = Value::new_instruction(
-            self.context,
-            self.block,
-            InstOp::ConditionalBranch {
-                cond_value,
-                true_block: BranchToWithArgs {
-                    block: true_block,
-                    args: true_dest_params,
-                },
-                false_block: BranchToWithArgs {
-                    block: false_block,
-                    args: false_dest_params,
-                },
-            },
-        );
-        true_block.add_pred(self.context, &self.block);
-        false_block.add_pred(self.context, &self.block);
-        self.context.blocks[self.block.0].instructions.push(cbr_val);
-        cbr_val
-    }
-
-    pub fn contract_call(
-        self,
-        return_type: Type,
-        name: String,
-        params: Value,
-        coins: Value,    // amount of coins to forward
-        asset_id: Value, // b256 asset ID of the coint being forwarded
-        gas: Value,      // amount of gas to forward
-    ) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::ContractCall {
-                return_type,
-                name,
-                params,
-                coins,
-                asset_id,
-                gas,
-            }
-        )
-    }
-
-    pub fn gtf(self, index: Value, tx_field_id: u64) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::Gtf { index, tx_field_id })
-        )
-    }
-
-    // get_elem_ptr() and get_elem_ptr_*() all take the element type and will store the pointer to
-    // that type in the instruction, which is later returned by Instruction::get_type().
-    pub fn get_elem_ptr(self, base: Value, elem_ty: Type, indices: Vec<Value>) -> Value {
-        let elem_ptr_ty = Type::new_ptr(self.context, elem_ty);
-        insert_instruction!(
-            self,
-            InstOp::GetElemPtr {
-                base,
-                elem_ptr_ty,
-                indices
-            }
-        )
-    }
-
-    pub fn get_elem_ptr_with_idx(self, base: Value, elem_ty: Type, index: u64) -> Value {
-        let idx_val = Constant::get_uint(self.context, 64, index);
-        self.get_elem_ptr(base, elem_ty, vec![idx_val])
-    }
-
-    pub fn get_elem_ptr_with_idcs(self, base: Value, elem_ty: Type, indices: &[u64]) -> Value {
-        let idx_vals = indices
-            .iter()
-            .map(|idx| Constant::get_uint(self.context, 64, *idx))
-            .collect();
-        self.get_elem_ptr(base, elem_ty, idx_vals)
-    }
-
-    pub fn get_local(self, local_var: LocalVar) -> Value {
-        insert_instruction!(self, InstOp::GetLocal(local_var))
-    }
-
-    pub fn int_to_ptr(self, value: Value, ty: Type) -> Value {
-        insert_instruction!(self, InstOp::IntToPtr(value, ty))
-    }
-
-    pub fn load(self, src_val: Value) -> Value {
-        insert_instruction!(self, InstOp::Load(src_val))
-    }
-
-    pub fn log(self, log_val: Value, log_ty: Type, log_id: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::Log {
-                log_val,
-                log_ty,
-                log_id
-            })
-        )
-    }
-
-    pub fn mem_copy_bytes(self, dst_val_ptr: Value, src_val_ptr: Value, byte_len: u64) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::MemCopyBytes {
-                dst_val_ptr,
-                src_val_ptr,
-                byte_len
-            }
-        )
-    }
-
-    pub fn mem_copy_val(self, dst_val_ptr: Value, src_val_ptr: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::MemCopyVal {
-                dst_val_ptr,
-                src_val_ptr,
-            }
-        )
-    }
-
-    pub fn nop(self) -> Value {
-        insert_instruction!(self, InstOp::Nop)
-    }
-
-    pub fn ptr_to_int(self, value: Value, ty: Type) -> Value {
-        insert_instruction!(self, InstOp::PtrToInt(value, ty))
-    }
-
-    pub fn read_register(self, reg: Register) -> Value {
-        insert_instruction!(self, InstOp::FuelVm(FuelVmInstruction::ReadRegister(reg)))
-    }
-
-    pub fn ret(self, value: Value, ty: Type) -> Value {
-        insert_instruction!(self, InstOp::Ret(value, ty))
-    }
-
-    pub fn revert(self, value: Value) -> Value {
-        let revert_val = Value::new_instruction(
-            self.context,
-            self.block,
-            InstOp::FuelVm(FuelVmInstruction::Revert(value)),
-        );
-        self.context.blocks[self.block.0]
-            .instructions
-            .push(revert_val);
-        revert_val
-    }
-
-    pub fn smo(self, recipient: Value, message: Value, message_size: Value, coins: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::Smo {
-                recipient,
-                message,
-                message_size,
-                coins,
-            })
-        )
-    }
-
-    pub fn state_clear(self, key: Value, number_of_slots: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::StateClear {
-                key,
-                number_of_slots
-            })
-        )
-    }
-
-    pub fn state_load_quad_word(
-        self,
-        load_val: Value,
-        key: Value,
-        number_of_slots: Value,
-    ) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::StateLoadQuadWord {
-                load_val,
-                key,
-                number_of_slots
-            })
-        )
-    }
-
-    pub fn state_load_word(self, key: Value) -> Value {
-        insert_instruction!(self, InstOp::FuelVm(FuelVmInstruction::StateLoadWord(key)))
-    }
-
-    pub fn state_store_quad_word(
-        self,
-        stored_val: Value,
-        key: Value,
-        number_of_slots: Value,
-    ) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::StateStoreQuadWord {
-                stored_val,
-                key,
-                number_of_slots
-            })
-        )
-    }
-
-    pub fn state_store_word(self, stored_val: Value, key: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::FuelVm(FuelVmInstruction::StateStoreWord { stored_val, key })
-        )
-    }
-
-    pub fn store(self, dst_val_ptr: Value, stored_val: Value) -> Value {
-        insert_instruction!(
-            self,
-            InstOp::Store {
-                dst_val_ptr,
-                stored_val,
-            }
-        )
     }
 }
